Correction long int.
Normalement avec ce lots de modifs, j'ai remplacé toutes les int apparentés à l'id par des long int.
Une nouvelle passe de correction est en cours.

+++ include/backtrack.h
@@ -6,5 +6,5 @@
 /* Algorithme de backtracking */
 int backtrack (Plateau *plateau,Carte tabCarte[],
 	       Chemin *parcours,
-	       int nombre_de_carte, int quiet, int une_sol);
+	       long int nombre_de_carte, int quiet, int une_sol);
 #endif
